refactor: 기물 구조 분리
