+++ src/main/java/chess/domain/piece/King.java
@@ -1,14 +1,14 @@
 package chess.domain.piece;
 
 import chess.domain.board.Board;
-import chess.domain.position.Position;
 import chess.domain.movepattern.KingMovePattern;
 import chess.domain.movepattern.MovePattern;
+import chess.domain.position.Position;
 import java.util.ArrayList;
 import java.util.Arrays;
 import java.util.List;
 
-public class King extends Piece {
+public class King extends ImmediatePiece {
 
     private final List<MovePattern> movePatterns;
 
@@ -21,33 +21,6 @@ public class King extends Piece {
     protected void validate(final Type type, final Side side) {
         validateType(type);
         validateSide(side);
-    }
-
-    @Override
-    public List<Position> findMovablePositions(final Position source, final Board board) {
-        final List<Position> movablePositions = new ArrayList<>();
-        final List<MovePattern> movePatterns = getMovePatterns();
-        for (MovePattern movePattern : movePatterns) {
-            Position nextPosition = source;
-            if (isRangeValid(nextPosition, movePattern)) {
-                nextPosition = nextPosition.move(movePattern);
-                checkSide(movablePositions, nextPosition, board);
-            }
-        }
-        return movablePositions;
-    }
-
-    private void checkSide(final List<Position> movablePositions, final Position nextPosition, final Board board) {
-        final Side nextSide = board.findSideByPosition(nextPosition);
-        if (nextSide != this.side) {
-            movablePositions.add(nextPosition);
-        }
-    }
-
-    private boolean isRangeValid(final Position position, final MovePattern movePattern) {
-        final int nextRank = position.getRankIndex() + movePattern.getRankVector();
-        final int nextFile = position.getFileIndex() + movePattern.getFileVector();
-        return nextRank >= 1 && nextRank <= 8 && nextFile >= 1 && nextFile <= 8;
     }
 
     @Override
