+++ src/main/java/chess/domain/piece/Piece.java
@@ -1,8 +1,8 @@
 package chess.domain.piece;
 
 import chess.domain.board.Board;
-import chess.domain.position.Position;
 import chess.domain.movepattern.MovePattern;
+import chess.domain.position.Position;
 import java.util.ArrayList;
 import java.util.List;
 
@@ -15,6 +15,12 @@ public abstract class Piece {
         validate(type, side);
         this.type = type;
         this.side = side;
+    }
+
+    protected boolean isRangeValid(final Position position, final MovePattern movePattern) {
+        final int nextRank = position.getRankIndex() + movePattern.getRankVector();
+        final int nextFile = position.getFileIndex() + movePattern.getFileVector();
+        return nextRank >= 1 && nextRank <= 8 && nextFile >= 1 && nextFile <= 8;
     }
 
     public boolean isPawn() {
@@ -37,31 +43,5 @@ public abstract class Piece {
 
     protected abstract List<MovePattern> getMovePatterns();
 
-    public List<Position> findMovablePositions(final Position source, final Board board) {
-        final List<Position> movablePositions = new ArrayList<>();
-        final List<MovePattern> movePatterns = getMovePatterns();
-        for (MovePattern movePattern : movePatterns) {
-            Position nextPosition = source;
-            boolean killFlag = false;
-            while (isRangeValid(nextPosition, movePattern) && !killFlag) {
-                nextPosition = nextPosition.move(movePattern);
-                killFlag = checkSide(movablePositions, nextPosition, board);
-            }
-        }
-        return movablePositions;
-    }
-
-    private boolean checkSide(final List<Position> movablePositions, final Position nextPosition, final Board board) {
-        final Side nextSide = board.findSideByPosition(nextPosition);
-        if (nextSide != this.side) {
-            movablePositions.add(nextPosition);
-        }
-        return nextSide != this.side && nextSide != Side.NEUTRALITY;
-    }
-
-    private boolean isRangeValid(final Position position, final MovePattern movePattern) {
-        final int nextRank = position.getRankIndex() + movePattern.getRankVector();
-        final int nextFile = position.getFileIndex() + movePattern.getFileVector();
-        return nextRank >= 1 && nextRank <= 8 && nextFile >= 1 && nextFile <= 8;
-    }
+    public abstract List<Position> findMovablePositions(final Position source, final Board board);
 }
